+++ R/ParameterSetCollection.R
@@ -197,6 +197,12 @@ ParameterSetCollection <- R6Class("ParameterSetCollection",
       newlst <- c(selflst, mlst)
       checkmate::assertNames(names(newlst), "strict")
       private$.parametersets <- newlst
+
+      selflst <- private$.supports
+      mlst <- data.table::rbindlist(lapply(lst, function(x) x$.__enclos_env__$private$.supports))
+      newlst <- rbind(selflst, mlst)
+      private$.supports <- newlst
+
       invisible(self)
     },
 
@@ -232,7 +238,7 @@ ParameterSetCollection <- R6Class("ParameterSetCollection",
 
   private = list(
     .parametersets = list(),
-    .supports = list(),
+    .supports = data.table(),
     .contains = function() {
       apply(private$.supports, 1, function(x) {
         assertContains(x[[2]], as.Tuple(unlist(self$getParameterValue(x[[1]]))),