+++ R/Wrapper_MixtureDistribution.R
@@ -1,7 +1,7 @@
-#' @name MixtureDistribution
 #' @title Mixture Distribution Wrapper
 #' @description Wrapper used to construct a mixture of two or more distributions.
 #'
+#' @template class_vecdist
 #' @template param_log
 #' @template param_logp
 #' @template param_simplify
@@ -9,11 +9,12 @@
 #' @template param_lowertail
 #' @template param_n
 #' @template param_decorators
-#' @template class_vecdist
 #'
-#' @return Returns an R6 object of class MixtureDistribution.
+#' @details A mixture distribution is defined by
 #'
-#' @seealso \code{\link{listWrappers}}
+#' \deqn{F_P(x) = w_1 F_{X1}(x) * ... * w_n F_{XN}(x)}{F_P(x) = w_1 F_X1(x) * ... * w_n F_XN(x)} #nolint
+#' where \eqn{F_P} is the cdf of the mixture distribution, \eqn{X1,...,XN} are
+#' independent distributions, and \eqn{w1,...,wN} are weights for the mixture.
 #'
 #' @export
 MixtureDistribution <- R6Class("MixtureDistribution",
@@ -33,9 +34,20 @@ MixtureDistribution <- R6Class("MixtureDistribution",
     initialize = function(distlist = NULL, weights = "uniform", distribution = NULL, params = NULL,
                           shared_params = NULL,
                           name = NULL, short_name = NULL,
-                          decorators = NULL) {
+                          decorators = NULL,
+                          vecdist = NULL) {
 
-      lng <- if (is.null(distlist)) length(params) else length(distlist)
+      if (!is.null(vecdist)) {
+        lng <- nrow(vecdist$modelTable)
+      } else if (!is.null(params)) {
+        if (checkmate::testList(params)) {
+          lng <- length(params)
+        } else {
+          lng <- nrow(params)
+        }
+      } else {
+        lng <- length(distlist)
+      }
 
       if (checkmate::testNumeric(weights)) {
         stopifnot(length(weights) == lng)
@@ -57,20 +69,50 @@ MixtureDistribution <- R6Class("MixtureDistribution",
         }
       )
 
-      super$initialize(
-        distlist = distlist,
-        distribution = distribution,
-        params = params,
-        shared_params = shared_params,
-        decorators = decorators,
-        outerID = "mix",
-        name = name,
-        short_name = short_name
-      )
+      if (!is.null(vecdist)) {
+        checkmate::assertClass(vecdist, "VectorDistribution")
 
-      if (!is.null(name)) self$name <- gsub("Vector", "Mixture", self$name)
-      if (!is.null(short_name)) self$short_name <- gsub("Vec", "Mix", self$short_name)
-      self$description <- gsub("Vector", "Mixture", self$description)
+        if (!is.null(decorators)) {
+          suppressMessages(decorate(self, decorators))
+        }
+
+        private$.modelTable <- vecdist$modelTable
+        private$.distlist <- vecdist$distlist
+        private$.univariate <- vecdist$.__enclos_env__$private$.univariate
+        private$.pdf <- vecdist$.__enclos_env__$private$.pdf
+        private$.cdf <- vecdist$.__enclos_env__$private$.cdf
+        private$.quantile <- vecdist$.__enclos_env__$private$.quantile
+        private$.rand <- vecdist$.__enclos_env__$private$.rand
+
+        super$.__enclos_env__$super$initialize(
+          distlist = if (vecdist$distlist) vecdist$wrappedModels() else NULL,
+          name = vecdist$name,
+          short_name = vecdist$short_name,
+          description = vecdist$description,
+          support = vecdist$properties$support,
+          type = vecdist$traits$type,
+          valueSupport = vecdist$traits$valueSupport,
+          variateForm = "multivariate",
+          parameters = vecdist$parameters(),
+          outerID = "mix"
+        )
+
+      } else {
+        super$initialize(
+          distlist = distlist,
+          distribution = distribution,
+          params = params,
+          shared_params = shared_params,
+          decorators = decorators,
+          outerID = "mix",
+          name = name,
+          short_name = short_name
+        )
+      }
+
+      if (is.null(name)) self$name <- gsub("Vector|Product", "Mixture", self$name)
+      if (is.null(short_name)) self$short_name <- gsub("Vec|Prod", "Mix", self$short_name)
+      self$description <- gsub("Vector|Product", "Mixture", self$description)
 
       if (private$.univariate) {
         private$.traits$variateForm <- "univariate"
@@ -201,5 +243,22 @@ MixtureDistribution <- R6Class("MixtureDistribution",
     .weights = numeric(0)
   )
 )
+
+#' @title Coercion to Mixture Distribution
+#' @description Helper functions to quickly convert compatible objects to
+#' a [MixtureDistribution].
+#' @param object [ProductDistribution] or [VectorDistribution]
+#' @param weights `(character(1)|numeric())`\cr
+#' Weights to use in the resulting mixture. If all distributions are weighted equally then
+#' `"uniform"` provides a much faster implementation, otherwise a vector of length equal
+#' to the number of wrapped distributions, this is automatically scaled internally.
+#' @export
+as.MixtureDistribution <- function(object, weights = "uniform") {
+  if (checkmate::testClass(object, "VectorDistribution")) {
+    return(MixtureDistribution$new(vecdist = object, weights = weights))
+  } else {
+    stop("Object must inherit from VectorDistribution.")
+  }
+}
 
 .distr6$wrappers <- append(.distr6$wrappers, list(MixtureDistribution = MixtureDistribution))