+++ tests/testthat/test-WrappedMixture.R
@@ -116,3 +116,19 @@ test_that("multivariate", {
                Multinomial$new(size = 8, probs = c(0.1, 0.9))$pdf(c(1, 2), c(7, 6)) * 0.1 +
                  Multinomial$new(size = 8, probs = c(0.3, 0.7))$pdf(c(1, 2), c(7, 6)) * 0.9)
 })
+
+test_that("vecdist constructor", {
+  v <- VectorDistribution$new(distribution = "Binom", params = data.frame(size = 1:2))
+  m <- MixtureDistribution$new(distribution = "Binom", params = data.frame(size = 1:2))
+  p <- ProductDistribution$new(distribution = "Binom", params = data.frame(size = 1:2))
+
+  expect_equal_distribution(as.MixtureDistribution(v), m)
+  expect_equal_distribution(as.MixtureDistribution(p), m)
+
+  v <- VectorDistribution$new(list(Binomial$new(size = 1), Binomial$new(size = 2)))
+  m <- MixtureDistribution$new(list(Binomial$new(size = 1), Binomial$new(size = 2)))
+  p <- ProductDistribution$new(list(Binomial$new(size = 1), Binomial$new(size = 2)))
+
+  expect_equal_distribution(as.MixtureDistribution(v), m)
+  expect_equal_distribution(as.MixtureDistribution(p), m)
+})